move click delta ui to the click delta module
